services/chat-agent/requirements.txt
# HTTP framework + ASGI server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Typed env-var loading
pydantic-settings==2.6.1

# Async SQLite driver
aiosqlite==0.20.0

# Async HTTP client (used to call DeepSeek)
httpx==0.27.2

# Vector database client (Qdrant)
qdrant-client==1.12.1

# Local embedding model client (Ollama)
ollama==0.4.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
